Rename InputData, OutputLocation => QueryBuffers, QueryBuffersMut and use buffer module
