+++ tiledb/api/src/query/buffer.rs
@@ -1,0 +1,129 @@
+use std::ops::{Deref, DerefMut};
+
+pub enum Buffer<'data, T = u8> {
+    Empty,
+    Borrowed(&'data [T]),
+    Owned(Box<[T]>),
+}
+
+impl<'data, T> Buffer<'data, T> {
+    pub fn size(&self) -> usize {
+        std::mem::size_of_val(self.as_ref())
+    }
+}
+
+impl<'data, T> AsRef<[T]> for Buffer<'data, T> {
+    fn as_ref(&self) -> &[T] {
+        match self {
+            Buffer::Empty => unsafe {
+                std::slice::from_raw_parts(
+                    std::ptr::NonNull::dangling().as_ptr(),
+                    0,
+                )
+            },
+            Buffer::Borrowed(data) => data,
+            Buffer::Owned(data) => data,
+        }
+    }
+}
+
+impl<'data, T> Deref for Buffer<'data, T> {
+    type Target = [T];
+    fn deref(&self) -> &Self::Target {
+        self.as_ref()
+    }
+}
+
+pub struct QueryBuffers<'data, T = u8> {
+    pub data: Buffer<'data, T>,
+    pub cell_offsets: Option<Buffer<'data, u64>>,
+}
+
+impl<'data, T> QueryBuffers<'data, T> {
+    pub fn borrow<'this>(&'this self) -> QueryBuffers<'data, T>
+    where
+        'this: 'data,
+    {
+        QueryBuffers {
+            data: Buffer::Borrowed(self.data.as_ref()),
+            cell_offsets: Option::map(self.cell_offsets.as_ref(), |c| {
+                Buffer::Borrowed(c.as_ref())
+            }),
+        }
+    }
+}
+
+pub enum BufferMut<'data, T = u8> {
+    Empty,
+    Borrowed(&'data mut [T]),
+    Owned(Box<[T]>),
+}
+
+impl<'data, T> BufferMut<'data, T> {
+    pub fn size(&self) -> usize {
+        std::mem::size_of_val(self.as_ref())
+    }
+}
+
+impl<'data, T> AsRef<[T]> for BufferMut<'data, T> {
+    fn as_ref(&self) -> &[T] {
+        match self {
+            BufferMut::Empty => unsafe {
+                std::slice::from_raw_parts(
+                    std::ptr::NonNull::dangling().as_ptr(),
+                    0,
+                )
+            },
+            BufferMut::Borrowed(data) => data,
+            BufferMut::Owned(data) => data,
+        }
+    }
+}
+
+impl<'data, T> AsMut<[T]> for BufferMut<'data, T> {
+    fn as_mut(&mut self) -> &mut [T] {
+        match self {
+            BufferMut::Empty => unsafe {
+                std::slice::from_raw_parts_mut(
+                    std::ptr::NonNull::dangling().as_ptr(),
+                    0,
+                )
+            },
+            BufferMut::Borrowed(data) => data,
+            BufferMut::Owned(data) => &mut *data,
+        }
+    }
+}
+
+impl<'data, T> Deref for BufferMut<'data, T> {
+    type Target = [T];
+    fn deref(&self) -> &Self::Target {
+        self.as_ref()
+    }
+}
+
+impl<'data, T> DerefMut for BufferMut<'data, T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        self.as_mut()
+    }
+}
+
+pub struct QueryBuffersMut<'data, T = u8> {
+    pub data: BufferMut<'data, T>,
+    pub cell_offsets: Option<BufferMut<'data, u64>>,
+}
+
+impl<'data, T> QueryBuffersMut<'data, T> {
+    /// Borrows this QueryBuffersMut to use as input data.
+    pub fn as_input<'this>(&'this self) -> QueryBuffers<'data, T>
+    where
+        'this: 'data,
+    {
+        QueryBuffers {
+            data: Buffer::Borrowed(self.data.as_ref()),
+            cell_offsets: Option::map(self.cell_offsets.as_ref(), |c| {
+                Buffer::Borrowed(c.as_ref())
+            }),
+        }
+    }
+}
